+++ src/main/java/com/github/davidmoten/rtree/geometry/Rectangle.java
@@ -2,22 +2,22 @@ package com.github.davidmoten.rtree.geometry;
 
 public interface Rectangle extends Geometry, HasGeometry {
 
-    float x1();
+    double x1();
 
-    float y1();
+    double y1();
 
-    float x2();
+    double x2();
 
-    float y2();
+    double y2();
 
-    float area();
+    double area();
 
     Rectangle add(Rectangle r);
 
     boolean contains(double x, double y);
 
-    float intersectionArea(Rectangle r);
+    double intersectionArea(Rectangle r);
 
-    float perimeter();
+    double perimeter();
 
 }
